Move unit tests to own tests_ file
