+++ python/dcr_template.py
@@ -30,16 +30,12 @@ from scipy.ndimage.interpolation import shift as scipy_shift
 import scipy.optimize.nnls as positive_lstsq
 
 import lsst.daf.persistence as daf_persistence
-import lsst.afw.coord as afwCoord
-import lsst.afw.geom as afwGeom
 import lsst.afw.image as afwImage
 import lsst.afw.math as afwMath
 import lsst.meas.algorithms as measAlg
 import lsst.pex.policy as pexPolicy
 from lsst.sims.photUtils import Bandpass, PhotometricParameters
 from lsst.utils import getPackageDir
-import unittest
-import lsst.utils.tests
 from .calc_refractive_index import diff_refraction
 
 __all__ = ["DcrModel", "DcrCorrection"]
@@ -92,18 +88,20 @@ class DcrModel:
             el = elevation_arr[img]
             az = azimuth_arr[img]
             pix = self.photoParams.platescale
-            dcr_gen = _dcr_generator(self.bandpass, pixel_scale=pix, elevation=el, azimuth=az)
+            dcr_gen = DcrModel._dcr_generator(self.bandpass, pixel_scale=pix, elevation=el, azimuth=az)
 
             if self.use_psf:
                 if use_full:
-                    dcr_kernel = _calc_psf_kernel_full(exp, dcr_gen, return_matrix=True, psf_img=self.psf_avg,
-                                                       x_size=self.kernel_size, y_size=self.kernel_size)
+                    dcr_kernel = DcrModel._calc_psf_kernel_full(exp, dcr_gen, x_size=self.kernel_size,
+                                                                y_size=self.kernel_size, psf_img=self.psf_avg,
+                                                                return_matrix=True)
                 else:
-                    dcr_kernel = _calc_psf_kernel(exp, dcr_gen, return_matrix=True, psf_img=self.psf_avg,
-                                                  x_size=self.kernel_size, y_size=self.kernel_size)
+                    dcr_kernel = DcrModel._calc_psf_kernel(exp, dcr_gen, x_size=self.kernel_size,
+                                                           y_size=self.kernel_size, psf_img=self.psf_avg,
+                                                           return_matrix=True)
             else:
-                dcr_kernel = _calc_offset_phase(exp, dcr_gen, return_matrix=True,
-                                                x_size=self.kernel_size, y_size=self.kernel_size)
+                dcr_kernel = DcrModel._calc_offset_phase(exp, dcr_gen, return_matrix=True,
+                                                         x_size=self.kernel_size, y_size=self.kernel_size)
             template = np.zeros((self.y_size, self.x_size))
             weights = np.zeros((self.y_size, self.x_size))
             pix_radius = self.kernel_size//2
@@ -183,6 +181,191 @@ class DcrModel:
             kernel = 1.0
         template[j - radius: j + radius + 1, i - radius: i + radius + 1] += vals*kernel
         weights[j - radius: j + radius + 1, i - radius: i + radius + 1] += kernel
+
+    # NOTE: This function was copied from StarFast.py
+    @staticmethod
+    def _load_bandpass(band_name='g', wavelength_step=None, use_mirror=True, use_lens=True, use_atmos=True,
+                       use_filter=True, use_detector=True, **kwargs):
+        """Load in Bandpass object from sims_photUtils."""
+        """
+        @param band_name: Common name of the filter used. For LSST, use u, g, r, i, z, or y
+        @param wavelength_step: Wavelength resolution, also the wavelength range of each sub-band plane
+        @param use_mirror: Flag, include mirror in filter throughput calculation?
+        @param use_lens: Flag, use LSST lens in filter throughput calculation?
+        @param use_atmos: Flag, use standard atmosphere transmission in filter throughput calculation?
+        @param use_filter: Flag, use LSST filters in filter throughput calculation?
+        """
+        class BandpassMod(Bandpass):
+            """Customize a few methods of the Bandpass class from sims_photUtils."""
+
+            def calc_eff_wavelen(self, wavelength_min=None, wavelength_max=None):
+                """Calculate effective wavelengths for filters."""
+                # This is useful for summary numbers for filters.
+                # Calculate effective wavelength of filters.
+                if self.phi is None:
+                    self.sbTophi()
+                if wavelength_min is None:
+                    wavelength_min = np.min(self.wavelen)
+                if wavelength_max is None:
+                    wavelength_max = np.max(self.wavelen)
+                w_inds = (self.wavelen >= wavelength_min) & (self.wavelen <= wavelength_max)
+                effwavelenphi = (self.wavelen[w_inds] * self.phi[w_inds]).sum() / self.phi[w_inds].sum()
+                return effwavelenphi
+
+            def calc_bandwidth(self):
+                f0 = constants.speed_of_light / (self.wavelen_min * 1.0e-9)
+                f1 = constants.speed_of_light / (self.wavelen_max * 1.0e-9)
+                f_cen = constants.speed_of_light / (self.calc_eff_wavelen() * 1.0e-9)
+                return(f_cen * 2.0 * (f0 - f1) / (f0 + f1))
+
+        """
+        Define the wavelength range and resolution for a given ugrizy band.
+        These are defined in case the LSST filter throughputs are not used.
+        """
+        band_dict = {'u': (324.0, 395.0), 'g': (405.0, 552.0), 'r': (552.0, 691.0),
+                     'i': (818.0, 921.0), 'z': (922.0, 997.0), 'y': (975.0, 1075.0)}
+        band_range = band_dict[band_name]
+        bandpass = BandpassMod(wavelen_min=band_range[0], wavelen_max=band_range[1],
+                               wavelen_step=wavelength_step)
+        throughput_dir = getPackageDir('throughputs')
+        lens_list = ['baseline/lens1.dat', 'baseline/lens2.dat', 'baseline/lens3.dat']
+        mirror_list = ['baseline/m1.dat', 'baseline/m2.dat', 'baseline/m3.dat']
+        atmos_list = ['atmos/atmos_11.dat']
+        detector_list = ['baseline/detector.dat']
+        filter_list = ['baseline/filter_' + band_name + '.dat']
+        component_list = []
+        if use_mirror:
+            component_list += mirror_list
+        if use_lens:
+            component_list += lens_list
+        if use_atmos:
+            component_list += atmos_list
+        if use_detector:
+            component_list += detector_list
+        if use_filter:
+            component_list += filter_list
+        bandpass.readThroughputList(rootDir=throughput_dir, componentList=component_list)
+        # Calculate bandpass phi value if required.
+        if bandpass.phi is None:
+            bandpass.sbTophi()
+        return(bandpass)
+
+    # NOTE: This function was copied from StarFast.py
+    @staticmethod
+    def _wavelength_iterator(bandpass, use_midpoint=False):
+        """Define iterator to ensure that loops over wavelength are consistent."""
+        wave_start = bandpass.wavelen_min
+        while np.ceil(wave_start) < bandpass.wavelen_max:
+            wave_end = wave_start + bandpass.wavelen_step
+            if wave_end > bandpass.wavelen_max:
+                wave_end = bandpass.wavelen_max
+            if use_midpoint:
+                yield(bandpass.calc_eff_wavelen(wavelength_min=wave_start, wavelength_max=wave_end))
+            else:
+                yield((wave_start, wave_end))
+            wave_start = wave_end
+
+    # NOTE: This function was modified from StarFast.py
+    @staticmethod
+    def _dcr_generator(bandpass, pixel_scale=None, elevation=50.0, azimuth=0.0, use_midpoint=False, **kwargs):
+        """Call the functions that compute Differential Chromatic Refraction (relative to mid-band)."""
+        """
+        @param bandpass: bandpass object created with load_bandpass
+        @param pixel_scale: plate scale in arcsec/pixel
+        @param elevation: elevation angle of the center of the image, in decimal degrees.
+        @param azimuth: azimuth angle of the observation, in decimal degrees.
+        """
+        zenith_angle = 90.0 - elevation
+        wavelength_midpoint = bandpass.calc_eff_wavelen()
+        delta = namedtuple("delta", ["start", "end"])
+        dcr = namedtuple("dcr", ["dx", "dy"])
+        for wl_start, wl_end in DcrModel._wavelength_iterator(bandpass, use_midpoint=False):
+            # Note that refract_amp can be negative, since it's relative to the midpoint of the full band
+            if use_midpoint:
+                wl_mid = bandpass.calc_eff_wavelen(wavelength_min=wl_start, wavelength_max=wl_end)
+                refract_mid = diff_refraction(wavelength=wl_mid, wavelength_ref=wavelength_midpoint,
+                                              zenith_angle=zenith_angle, **kwargs)
+                refract_mid *= 3600.0 / pixel_scale
+                dx_mid = refract_mid * np.sin(np.radians(azimuth))
+                dy_mid = refract_mid * np.cos(np.radians(azimuth))
+                yield(dcr(dx=dx_mid, dy=dy_mid))
+            else:
+                refract_start = diff_refraction(wavelength=wl_start, wavelength_ref=wavelength_midpoint,
+                                                zenith_angle=zenith_angle, **kwargs)
+                refract_end = diff_refraction(wavelength=wl_end, wavelength_ref=wavelength_midpoint,
+                                              zenith_angle=zenith_angle, **kwargs)
+                refract_start *= 3600.0 / pixel_scale  # Refraction initially in degrees, convert to pixels.
+                refract_end *= 3600.0 / pixel_scale
+                dx = delta(start=refract_start * np.sin(np.radians(azimuth)),
+                           end=refract_end * np.sin(np.radians(azimuth)))
+                dy = delta(start=refract_start * np.cos(np.radians(azimuth)),
+                           end=refract_end * np.cos(np.radians(azimuth)))
+                yield(dcr(dx=dx, dy=dy))
+
+    @staticmethod
+    def _calc_offset_phase(exposure, dcr_gen, x_size=None, y_size=None, return_matrix=False):
+        """Return the 2D FFT of an offset generated by _dcr_generator in the form (dx, dy)."""
+        phase_arr = []
+        if y_size is None:
+            y_size = exposure.getHeight()
+        if x_size is None:
+            x_size = exposure.getWidth()
+        for dx, dy in dcr_gen:
+            kernel_x = _kernel_1d(dx, x_size)
+            kernel_y = _kernel_1d(dy, y_size)
+            kernel = np.einsum('i,j->ij', kernel_y, kernel_x)
+            if return_matrix:
+                shift_mat = np.zeros((x_size*y_size, x_size*y_size))
+                for j in range(y_size):
+                    for i in range(x_size):
+                        ij = i + j * x_size
+                        shift_mat[ij, :] = np.ravel(scipy_shift(kernel, (j - y_size//2, i - x_size//2),
+                                                    mode='constant', cval=0.0))
+                phase_arr.append(shift_mat)
+            else:
+                phase_arr.append(np.ravel(kernel))
+        phase_arr = np.vstack(phase_arr)
+        return(phase_arr)
+
+    @staticmethod
+    def _calc_psf_kernel(exposure, dcr_gen, x_size=None, y_size=None, return_matrix=False, psf_img=None):
+
+        if y_size is None:
+            y_size = exposure.getHeight()
+        if x_size is None:
+            x_size = exposure.getWidth()
+        psf_kernel_arr = []
+        for dcr in dcr_gen:
+            psf_y_size, psf_x_size = psf_img.shape
+            psf = np.zeros((y_size, x_size), dtype=psf_img.dtype)
+            if return_matrix:
+                psf_kernel_arr.append(_calc_psf_kernel_subroutine(psf_img, dcr))
+            else:
+                psf_kernel_arr.append(np.ravel(psf))
+
+        psf_kernel_arr = np.vstack(psf_kernel_arr)
+        return(psf_kernel_arr)
+
+    @staticmethod
+    def _calc_psf_kernel_full(exposure, dcr_gen, x_size=None, y_size=None, return_matrix=False, psf_img=None):
+        # psf_img is passed in but not used so that this function can be swapped in for _calc_psf_kernel
+        # without having to change the keywords
+        if y_size is None:
+            y_size = exposure.getHeight()
+        if x_size is None:
+            x_size = exposure.getWidth()
+        psf_kernel_arr = []
+        psf_img = exposure.getPsf().computeKernelImage().getArray()
+        for dcr in dcr_gen:
+            kernel_single = _calc_psf_kernel_subroutine(psf_img, dcr, x_size=x_size, y_size=y_size)
+            if return_matrix:
+                psf_kernel_arr.append(kernel_single)
+            else:
+                kernel_single = kernel_single[y_size*x_size//2, :]
+                psf_kernel_arr.append(kernel_single)
+
+        psf_kernel_arr = np.vstack(psf_kernel_arr)
+        return(psf_kernel_arr)
 
     def _edge_test(self, j, i):
         x0 = 150
@@ -266,7 +449,7 @@ class DcrModel:
             butler = self.butler
         else:
             butler = daf_persistence.Butler(model_repository)
-        wave_gen = _wavelength_iterator(self.bandpass, use_midpoint=False)
+        wave_gen = DcrModel._wavelength_iterator(self.bandpass, use_midpoint=False)
         for f in range(self.n_step):
             wl_start, wl_end = wave_gen.next()
             exp = self._create_exposure(self.model[f, :, :], variance=self.weights[f, :, :],
@@ -297,7 +480,7 @@ class DcrModel:
         meta = dcrModel.getMetadata()
         self.wcs = dcrModel.getWcs()
         self.n_step = len(model_arr)
-        wavelength_step = meta.get("WAVESTEP")
+        wave_step = meta.get("WAVESTEP")
         self.use_psf = meta.get("PSF_FLAG")
         self.y_size, self.x_size = dcrModel.getDimensions()
         self.pixel_scale = self.wcs.pixelScale().asArcseconds()
@@ -306,7 +489,7 @@ class DcrModel:
                                                  bandpass=band_name)
         self.bbox = dcrModel.getBBox()
         self.kernel_size = meta.get("KSUPPORT")
-        self.bandpass = _load_bandpass(band_name=band_name, wavelength_step=wavelength_step, **kwargs)
+        self.bandpass = DcrModel._load_bandpass(band_name=band_name, wavelength_step=wave_step, **kwargs)
 
         self.psf = dcrModel.getPsf()
         self.psf_size = self.psf.computeKernelImage().getArray().shape[0]
@@ -368,16 +551,16 @@ class DcrCorrection(DcrModel):
         self.psf_size = calexp.getPsf().computeKernelImage().getArray().shape[0]
         self._combine_masks()
 
-        bandpass = _load_bandpass(band_name=band_name, wavelength_step=wavelength_step, **kwargs)
+        bandpass = DcrModel._load_bandpass(band_name=band_name, wavelength_step=wavelength_step, **kwargs)
         if n_step is not None:
             wavelength_step = (bandpass.wavelen_max - bandpass.wavelen_min) / n_step
-            bandpass = _load_bandpass(band_name=band_name, wavelength_step=wavelength_step, **kwargs)
+            bandpass = DcrModel._load_bandpass(band_name=band_name, wavelength_step=wavelength_step, **kwargs)
         else:
             n_step = int(np.ceil((bandpass.wavelen_max - bandpass.wavelen_min) / bandpass.wavelen_step))
         if n_step >= self.n_images:
             print("Warning! Under-constrained system. Reducing number of frequency planes.")
             wavelength_step *= n_step / self.n_images
-            bandpass = _load_bandpass(band_name=band_name, wavelength_step=wavelength_step, **kwargs)
+            bandpass = DcrModel._load_bandpass(band_name=band_name, wavelength_step=wavelength_step, **kwargs)
             n_step = int(np.ceil((bandpass.wavelen_max - bandpass.wavelen_min) / bandpass.wavelen_step))
         self.n_step = n_step
         self.bandpass = bandpass
@@ -385,7 +568,8 @@ class DcrCorrection(DcrModel):
                                                  bandpass=band_name)
 
         elevation_min = np.min(self.elevation_arr) - 5.  # Calculate slightly worse DCR than maximum.
-        dcr_test = _dcr_generator(bandpass, pixel_scale=self.pixel_scale, elevation=elevation_min, azimuth=0.)
+        dcr_test = DcrModel._dcr_generator(bandpass, pixel_scale=self.pixel_scale,
+                                           elevation=elevation_min, azimuth=0.)
         self.dcr_max = int(np.ceil(np.max(dcr_test.next())) + 1)
         if kernel_size is None:
             self.kernel_size = 2*self.dcr_max + 1
@@ -410,11 +594,6 @@ class DcrCorrection(DcrModel):
         reg_lambda2 = None
 
         if spatial_regularization:
-            # baseReg = _difference(kernel_size)
-            # reg_pix = np.zeros((n_step * kernel_size, n_step * kernel_size))
-
-            # for i in range(n_step):
-            #     reg_pix[i::n_step, i::n_step] = baseReg
             reg_pix_x = np.zeros((n_step * x_size * y_size,
                                   n_step * x_size * y_size - x_size))
             for ij in range(n_step * x_size * y_size - x_size):
@@ -489,12 +668,14 @@ class DcrCorrection(DcrModel):
 
             # Use the measured PSF as the solution of the shifted PSFs.
             # Taken at zenith, since we're solving for the shift and don't want to introduce any extra.
-            dcr_genZ = _dcr_generator(self.bandpass, pixel_scale=self.pixel_scale, elevation=90., azimuth=az)
-            psf_mat.append(_calc_psf_kernel_full(exp, dcr_genZ, return_matrix=False,
+            dcr_genZ = DcrModel._dcr_generator(self.bandpass, pixel_scale=self.pixel_scale,
+                                               elevation=90., azimuth=az)
+            psf_mat.append(DcrModel._calc_psf_kernel_full(exp, dcr_genZ, return_matrix=False,
                            x_size=self.psf_size, y_size=self.psf_size))
             # Calculate the expected shift (with no psf) due to DCR
-            dcr_gen = _dcr_generator(self.bandpass, pixel_scale=self.pixel_scale, elevation=el, azimuth=az)
-            dcr_shift.append(_calc_offset_phase(exp, dcr_gen, return_matrix=True,
+            dcr_gen = DcrModel._dcr_generator(self.bandpass, pixel_scale=self.pixel_scale,
+                                              elevation=el, azimuth=az)
+            dcr_shift.append(DcrModel._calc_offset_phase(exp, dcr_gen, return_matrix=True,
                              x_size=self.psf_size, y_size=self.psf_size))
         psf_mat = np.sum(np.hstack(psf_mat), axis=0)
         dcr_shift = np.hstack(dcr_shift)
@@ -556,17 +737,18 @@ class DcrCorrection(DcrModel):
         for img, exp in enumerate(self.exposures):
             el = self.elevation_arr[img]
             az = self.azimuth_arr[img]
-            dcr_gen = _dcr_generator(self.bandpass, pixel_scale=self.pixel_scale, elevation=el, azimuth=az)
+            dcr_gen = DcrModel._dcr_generator(self.bandpass, pixel_scale=self.pixel_scale,
+                                              elevation=el, azimuth=az)
 
             if self.use_psf:
                 if use_full:
-                    dcr_kernel.append(_calc_psf_kernel_full(exp, dcr_gen, psf_img=self.psf_avg,
+                    dcr_kernel.append(DcrModel._calc_psf_kernel_full(exp, dcr_gen, psf_img=self.psf_avg,
                                       x_size=self.kernel_size, y_size=self.kernel_size, return_matrix=True))
                 else:
-                    dcr_kernel.append(_calc_psf_kernel(exp, dcr_gen, psf_img=self.psf_avg,
+                    dcr_kernel.append(DcrModel._calc_psf_kernel(exp, dcr_gen, psf_img=self.psf_avg,
                                       x_size=self.kernel_size, y_size=self.kernel_size, return_matrix=True))
             else:
-                dcr_kernel.append(_calc_offset_phase(exp, dcr_gen, return_matrix=True,
+                dcr_kernel.append(DcrModel._calc_offset_phase(exp, dcr_gen, return_matrix=True,
                                   x_size=self.kernel_size, y_size=self.kernel_size))
         dcr_kernel = np.hstack(dcr_kernel)
         return(dcr_kernel)
@@ -591,71 +773,6 @@ class DcrCorrection(DcrModel):
         model_solution = positive_lstsq(kernel_use, vals_use)
         model_vals = model_solution[0]
         return(np.reshape(model_vals, (self.n_step, y_size, x_size)))
-
-
-def _calc_offset_phase(exposure, dcr_gen, x_size=None, y_size=None, return_matrix=False):
-    """Return the 2D FFT of an offset generated by _dcr_generator in the form (dx, dy)."""
-    phase_arr = []
-    if y_size is None:
-        y_size = exposure.getHeight()
-    if x_size is None:
-        x_size = exposure.getWidth()
-    for dx, dy in dcr_gen:
-        kernel_x = _kernel_1d(dx, x_size)
-        kernel_y = _kernel_1d(dy, y_size)
-        kernel = np.einsum('i,j->ij', kernel_y, kernel_x)
-        if return_matrix:
-            shift_mat = np.zeros((x_size*y_size, x_size*y_size))
-            for j in range(y_size):
-                for i in range(x_size):
-                    ij = i + j * x_size
-                    shift_mat[ij, :] = np.ravel(scipy_shift(kernel, (j - y_size//2, i - x_size//2),
-                                                mode='constant', cval=0.0))
-            phase_arr.append(shift_mat)
-        else:
-            phase_arr.append(np.ravel(kernel))
-    phase_arr = np.vstack(phase_arr)
-    return(phase_arr)
-
-
-def _calc_psf_kernel(exposure, dcr_gen, x_size=None, y_size=None, return_matrix=False, psf_img=None):
-
-    if y_size is None:
-        y_size = exposure.getHeight()
-    if x_size is None:
-        x_size = exposure.getWidth()
-    psf_kernel_arr = []
-    for dcr in dcr_gen:
-        psf_y_size, psf_x_size = psf_img.shape
-        psf = np.zeros((y_size, x_size), dtype=psf_img.dtype)
-        if return_matrix:
-            psf_kernel_arr.append(_calc_psf_kernel_subroutine(psf_img, dcr))
-        else:
-            psf_kernel_arr.append(np.ravel(psf))
-
-    psf_kernel_arr = np.vstack(psf_kernel_arr)
-    return(psf_kernel_arr)
-
-
-def _calc_psf_kernel_full(exposure, dcr_gen, x_size=None, y_size=None, return_matrix=False, psf_img=None):
-    # psf_img is passed in but not used so that this function can be swapped in for _calc_psf_kernel
-    # without having to change the keywords
-    if y_size is None:
-        y_size = exposure.getHeight()
-    if x_size is None:
-        x_size = exposure.getWidth()
-    psf_kernel_arr = []
-    psf_img = exposure.getPsf().computeKernelImage().getArray()
-    for dcr in dcr_gen:
-        kernel_single = _calc_psf_kernel_subroutine(psf_img, dcr, x_size=x_size, y_size=y_size)
-        if return_matrix:
-            psf_kernel_arr.append(kernel_single)
-        else:
-            kernel_single = kernel_single[y_size*x_size//2, :]
-            psf_kernel_arr.append(kernel_single)
-
-    psf_kernel_arr = np.vstack(psf_kernel_arr)
-    return(psf_kernel_arr)
 
 
 def _calc_psf_kernel_subroutine(psf_img, dcr, x_size=None, y_size=None):
@@ -689,126 +806,6 @@ def _calc_psf_kernel_subroutine(psf_img, dcr, x_size=None, y_size=None):
     return(psf_mat)
 
 
-# NOTE: This function was copied from StarFast.py
-def _load_bandpass(band_name='g', wavelength_step=None, use_mirror=True, use_lens=True, use_atmos=True,
-                   use_filter=True, use_detector=True, **kwargs):
-    """Load in Bandpass object from sims_photUtils."""
-    """
-    @param band_name: Common name of the filter used. For LSST, use u, g, r, i, z, or y
-    @param wavelength_step: Wavelength resolution, also the wavelength range of each sub-band plane
-    @param use_mirror: Flag, include mirror in filter throughput calculation?
-    @param use_lens: Flag, use LSST lens in filter throughput calculation?
-    @param use_atmos: Flag, use standard atmosphere transmission in filter throughput calculation?
-    @param use_filter: Flag, use LSST filters in filter throughput calculation?
-    """
-    class BandpassMod(Bandpass):
-        """Customize a few methods of the Bandpass class from sims_photUtils."""
-
-        def calc_eff_wavelen(self, wavelength_min=None, wavelength_max=None):
-            """Calculate effective wavelengths for filters."""
-            # This is useful for summary numbers for filters.
-            # Calculate effective wavelength of filters.
-            if self.phi is None:
-                self.sbTophi()
-            if wavelength_min is None:
-                wavelength_min = np.min(self.wavelen)
-            if wavelength_max is None:
-                wavelength_max = np.max(self.wavelen)
-            w_inds = (self.wavelen >= wavelength_min) & (self.wavelen <= wavelength_max)
-            effwavelenphi = (self.wavelen[w_inds] * self.phi[w_inds]).sum() / self.phi[w_inds].sum()
-            return effwavelenphi
-
-        def calc_bandwidth(self):
-            f0 = constants.speed_of_light / (self.wavelen_min * 1.0e-9)
-            f1 = constants.speed_of_light / (self.wavelen_max * 1.0e-9)
-            f_cen = constants.speed_of_light / (self.calc_eff_wavelen() * 1.0e-9)
-            return(f_cen * 2.0 * (f0 - f1) / (f0 + f1))
-
-    """
-    Define the wavelength range and resolution for a given ugrizy band.
-    These are defined in case the LSST filter throughputs are not used.
-    """
-    band_dict = {'u': (324.0, 395.0), 'g': (405.0, 552.0), 'r': (552.0, 691.0),
-                 'i': (818.0, 921.0), 'z': (922.0, 997.0), 'y': (975.0, 1075.0)}
-    band_range = band_dict[band_name]
-    bandpass = BandpassMod(wavelen_min=band_range[0], wavelen_max=band_range[1],
-                           wavelen_step=wavelength_step)
-    throughput_dir = getPackageDir('throughputs')
-    lens_list = ['baseline/lens1.dat', 'baseline/lens2.dat', 'baseline/lens3.dat']
-    mirror_list = ['baseline/m1.dat', 'baseline/m2.dat', 'baseline/m3.dat']
-    atmos_list = ['atmos/atmos_11.dat']
-    detector_list = ['baseline/detector.dat']
-    filter_list = ['baseline/filter_' + band_name + '.dat']
-    component_list = []
-    if use_mirror:
-        component_list += mirror_list
-    if use_lens:
-        component_list += lens_list
-    if use_atmos:
-        component_list += atmos_list
-    if use_detector:
-        component_list += detector_list
-    if use_filter:
-        component_list += filter_list
-    bandpass.readThroughputList(rootDir=throughput_dir, componentList=component_list)
-    # Calculate bandpass phi value if required.
-    if bandpass.phi is None:
-        bandpass.sbTophi()
-    return(bandpass)
-
-
-# NOTE: This function was copied from StarFast.py
-def _wavelength_iterator(bandpass, use_midpoint=False):
-    """Define iterator to ensure that loops over wavelength are consistent."""
-    wave_start = bandpass.wavelen_min
-    while np.ceil(wave_start) < bandpass.wavelen_max:
-        wave_end = wave_start + bandpass.wavelen_step
-        if wave_end > bandpass.wavelen_max:
-            wave_end = bandpass.wavelen_max
-        if use_midpoint:
-            yield(bandpass.calc_eff_wavelen(wavelength_min=wave_start, wavelength_max=wave_end))
-        else:
-            yield((wave_start, wave_end))
-        wave_start = wave_end
-
-
-# NOTE: This function was modified from StarFast.py
-def _dcr_generator(bandpass, pixel_scale=None, elevation=50.0, azimuth=0.0, use_midpoint=False, **kwargs):
-    """Call the functions that compute Differential Chromatic Refraction (relative to mid-band)."""
-    """
-    @param bandpass: bandpass object created with load_bandpass
-    @param pixel_scale: plate scale in arcsec/pixel
-    @param elevation: elevation angle of the center of the image, in decimal degrees.
-    @param azimuth: azimuth angle of the observation, in decimal degrees.
-    """
-    zenith_angle = 90.0 - elevation
-    wavelength_midpoint = bandpass.calc_eff_wavelen()
-    delta = namedtuple("delta", ["start", "end"])
-    dcr = namedtuple("dcr", ["dx", "dy"])
-    for wl_start, wl_end in _wavelength_iterator(bandpass, use_midpoint=False):
-        # Note that refract_amp can be negative, since it's relative to the midpoint of the full band
-        if use_midpoint:
-            wl_mid = bandpass.calc_eff_wavelen(wavelength_min=wl_start, wavelength_max=wl_end)
-            refract_mid = diff_refraction(wavelength=wl_mid, wavelength_ref=wavelength_midpoint,
-                                          zenith_angle=zenith_angle, **kwargs)
-            refract_mid *= 3600.0 / pixel_scale
-            dx_mid = refract_mid * np.sin(np.radians(azimuth))
-            dy_mid = refract_mid * np.cos(np.radians(azimuth))
-            yield(dcr(dx=dx_mid, dy=dy_mid))
-        else:
-            refract_start = diff_refraction(wavelength=wl_start, wavelength_ref=wavelength_midpoint,
-                                            zenith_angle=zenith_angle, **kwargs)
-            refract_end = diff_refraction(wavelength=wl_end, wavelength_ref=wavelength_midpoint,
-                                          zenith_angle=zenith_angle, **kwargs)
-            refract_start *= 3600.0 / pixel_scale  # Refraction initially in degrees, convert to pixels.
-            refract_end *= 3600.0 / pixel_scale
-            dx = delta(start=refract_start * np.sin(np.radians(azimuth)),
-                       end=refract_end * np.sin(np.radians(azimuth)))
-            dy = delta(start=refract_start * np.cos(np.radians(azimuth)),
-                       end=refract_end * np.cos(np.radians(azimuth)))
-            yield(dcr(dx=dx, dy=dy))
-
-
 def _kernel_1d(offset, size, width=0.0, min_width=0.1):
     """Pre-compute the 1D sinc function values along each axis."""
     """
@@ -832,570 +829,3 @@ def _kernel_1d(offset, size, width=0.0, min_width=0.1):
 class NonnegLstsqIterFit():
     def __init__():
         pass
-
-# -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
-
-
-class _BasicBandpass:
-    """Dummy bandpass object for testing."""
-
-    def __init__(self, band_name='g', wavelength_step=1):
-        """Define the wavelength range and resolution for a given ugrizy band."""
-        band_dict = {'u': (324.0, 395.0), 'g': (405.0, 552.0), 'r': (552.0, 691.0),
-                     'i': (818.0, 921.0), 'z': (922.0, 997.0), 'y': (975.0, 1075.0)}
-        band_range = band_dict[band_name]
-        self.wavelen_min = band_range[0]
-        self.wavelen_max = band_range[1]
-        self.wavelen_step = wavelength_step
-
-    def calc_eff_wavelen(self, wavelength_min=None, wavelength_max=None):
-        """Mimic the calc_eff_wavelen method of the real bandpass class."""
-        if wavelength_min is None:
-            wavelength_min = self.wavelen_min
-        if wavelength_max is None:
-            wavelength_max = self.wavelen_max
-        return((wavelength_min + wavelength_max) / 2.0)
-
-    def calc_bandwidth(self):
-        f0 = constants.speed_of_light / (self.wavelen_min * 1.0e-9)
-        f1 = constants.speed_of_light / (self.wavelen_max * 1.0e-9)
-        f_cen = constants.speed_of_light / (self.calc_eff_wavelen() * 1.0e-9)
-        return(f_cen * 2.0 * (f0 - f1) / (f0 + f1))
-
-    def getBandpass(self):
-        """Mimic the getBandpass method of the real bandpass class."""
-        wl_gen = _wavelength_iterator(self)
-        wavelengths = [wl[0] for wl in wl_gen]
-        wavelengths += [self.wavelen_max]
-        bp_vals = [1] * len(wavelengths)
-        return((wavelengths, bp_vals))
-
-
-def _create_wcs(bbox=None, pixel_scale=None, ra=None, dec=None, sky_rotation=None):
-    """Create a wcs (coordinate system)."""
-    crval = afwCoord.IcrsCoord(ra * afwGeom.degrees, dec * afwGeom.degrees)
-    crpix = afwGeom.Box2D(bbox).getCenter()
-    cd1_1 = (pixel_scale * afwGeom.arcseconds * np.cos(np.radians(sky_rotation))).asDegrees()
-    cd1_2 = (-pixel_scale * afwGeom.arcseconds * np.sin(np.radians(sky_rotation))).asDegrees()
-    cd2_1 = (pixel_scale * afwGeom.arcseconds * np.sin(np.radians(sky_rotation))).asDegrees()
-    cd2_2 = (pixel_scale * afwGeom.arcseconds * np.cos(np.radians(sky_rotation))).asDegrees()
-    return(afwImage.makeWcs(crval, crpix, cd1_1, cd1_2, cd2_1, cd2_2))
-
-
-class _BasicDcrModel(DcrModel):
-    """Dummy DcrModel object for testing without a repository."""
-
-    def __init__(self, size=None, kernel_size=5, n_step=3, band_name='g', exposure_time=30.,
-                 pixel_scale=0.25, wavelength_step=10.0):
-        seed = 5
-        rand_gen = np.random
-        rand_gen.seed(seed)
-        self.butler = None
-        self.use_psf = False
-
-        bandpass_init = _BasicBandpass(band_name=band_name, wavelength_step=None)
-        wavelength_step = (bandpass_init.wavelen_max - bandpass_init.wavelen_min) / n_step
-        self.bandpass = _BasicBandpass(band_name=band_name, wavelength_step=wavelength_step)
-        self.model = rand_gen.random(size=(n_step, size, size))
-        self.weights = np.ones((n_step, size, size))
-        self.mask = np.zeros((size, size), dtype=np.int32)
-
-        self.n_step = n_step
-        self.y_size = size
-        self.x_size = size
-        self.pixel_scale = pixel_scale
-        self.kernel_size = kernel_size
-        self.photoParams = PhotometricParameters(exptime=exposure_time, nexp=1, platescale=pixel_scale,
-                                                 bandpass=band_name)
-        self.bbox = afwGeom.Box2I(afwGeom.Point2I(0, 0), afwGeom.ExtentI(size, size))
-        self.wcs = _create_wcs(bbox=self.bbox, pixel_scale=pixel_scale, ra=0., dec=0., sky_rotation=0.)
-
-        psf_vals = np.zeros((kernel_size, kernel_size))
-        psf_vals[kernel_size//2 - 1: kernel_size//2 + 1,
-                 kernel_size//2 - 1: kernel_size//2 + 1] = 0.5
-        psf_vals[kernel_size//2, kernel_size//2] = 1.
-        psf_image = afwImage.ImageD(kernel_size, kernel_size)
-        psf_image.getArray()[:, :] = psf_vals
-        psfK = afwMath.FixedKernel(psf_image)
-        self.psf = measAlg.KernelPsf(psfK)
-
-        self.psf_avg = psf_vals  # self.psf.computeKernelImage().getArray()
-
-
-class _BasicDcrCorrection(DcrCorrection):
-    """Dummy DcrCorrection object for testing without a repository."""
-
-    def __init__(self, band_name='g', n_step=3, use_psf=False, kernel_size=5, exposures=None):
-        self.butler = None
-        self.use_psf = use_psf
-        self.debug = False
-
-        self.elevation_arr = []
-        self.azimuth_arr = []
-        self.airmass_arr = []
-        for calexp in exposures:
-            self.elevation_arr.append(90 - calexp.getMetadata().get("ZENITH"))
-            self.azimuth_arr.append(calexp.getMetadata().get("AZIMUTH"))
-            self.airmass_arr.append(calexp.getMetadata().get("AIRMASS"))
-        self.exposures = exposures
-
-        bandpass_init = _BasicBandpass(band_name=band_name, wavelength_step=None)
-        wavelength_step = (bandpass_init.wavelen_max - bandpass_init.wavelen_min) / n_step
-        self.bandpass = _BasicBandpass(band_name=band_name, wavelength_step=wavelength_step)
-        self.n_step = n_step
-        self.n_images = len(exposures)
-        self.y_size, self.x_size = exposures[0].getDimensions()
-        self.pixel_scale = calexp.getWcs().pixelScale().asArcseconds()
-        exposure_time = calexp.getInfo().getCalib().getExptime()
-        self.bbox = calexp.getBBox()
-        self.wcs = calexp.getWcs()
-        self.psf_size = calexp.getPsf().computeKernelImage().getArray().shape[0]
-        self.kernel_size = kernel_size
-        self.photoParams = PhotometricParameters(exptime=exposure_time, nexp=1, platescale=self.pixel_scale,
-                                                 bandpass=band_name)
-        elevation_min = np.min(self.elevation_arr) - 5.  # Calculate slightly worse DCR than maximum.
-        dcr_test = _dcr_generator(self.bandpass, pixel_scale=self.pixel_scale,
-                                  elevation=elevation_min, azimuth=0.)
-        self.dcr_max = int(np.ceil(np.max(dcr_test.next())) + 1)
-        if kernel_size is None:
-            self.kernel_size = 2*self.dcr_max + 1
-        else:
-            self.kernel_size = kernel_size
-
-
-class DCRTestCase(lsst.utils.tests.TestCase):
-    """Test the the calculations of Differential Chromatic Refraction."""
-
-    def setUp(self):
-        """Define parameters used by every test."""
-        band_name = 'g'
-        wavelength_step = 10.0
-        self.pixel_scale = 0.25
-        self.bandpass = _BasicBandpass(band_name=band_name, wavelength_step=wavelength_step)
-
-    def tearDown(self):
-        """Clean up."""
-        del self.bandpass
-
-    def test_dcr_generator(self):
-        """Check that _dcr_generator returns a generator with n_step iterations, and (0,0) at zenith."""
-        azimuth = 0.0
-        elevation = 90.0
-        zenith_dcr = 0.
-        bp = self.bandpass
-        dcr_gen = _dcr_generator(bp, pixel_scale=self.pixel_scale, elevation=elevation, azimuth=azimuth)
-        n_step = int(np.ceil((bp.wavelen_max - bp.wavelen_min) / bp.wavelen_step))
-        for f in range(n_step):
-            dcr = next(dcr_gen)
-            self.assertFloatsEqual(dcr.dx.start, zenith_dcr)
-            self.assertFloatsEqual(dcr.dx.end, zenith_dcr)
-            self.assertFloatsEqual(dcr.dy.start, zenith_dcr)
-            self.assertFloatsEqual(dcr.dy.end, zenith_dcr)
-        # Also check that the generator is now exhausted
-        with self.assertRaises(StopIteration):
-            next(dcr_gen)
-
-    def test_dcr_values(self):
-        """Check DCR against pre-computed values."""
-        azimuth = 0.0
-        elevation = 50.0
-        dcr_ref_vals = [(1.9847367904770623, 1.6467981843302726),
-                        (1.6467981843302726, 1.3341803407311699),
-                        (1.3341803407311699, 1.0443731947908652),
-                        (1.0443731947908652, 0.77517513542339489),
-                        (0.77517513542339489, 0.52464791969238367),
-                        (0.52464791969238367, 0.29107920440002155),
-                        (0.29107920440002155, 0.072951238300825172),
-                        (0.072951238300825172, -0.13108543143740825),
-                        (-0.13108543143740825, -0.3222341473268886),
-                        (-0.3222341473268886, -0.50157094957602733),
-                        (-0.50157094957602733, -0.6700605866796161),
-                        (-0.6700605866796161, -0.8285701993878597),
-                        (-0.8285701993878597, -0.97788106062563773),
-                        (-0.97788106062563773, -1.1186986838806061),
-                        (-1.1186986838806061, -1.2125619138571659)]
-        bp = self.bandpass
-        dcr_gen = _dcr_generator(bp, pixel_scale=self.pixel_scale, elevation=elevation,
-                                 azimuth=azimuth, use_midpoint=False)
-        n_step = int(np.ceil((bp.wavelen_max - bp.wavelen_min) / bp.wavelen_step))
-        for f in range(n_step):
-            dcr = next(dcr_gen)
-            self.assertFloatsEqual(dcr.dy.start, dcr_ref_vals[f][0])
-            self.assertFloatsEqual(dcr.dy.end, dcr_ref_vals[f][1])
-
-
-class BandpassTestCase(lsst.utils.tests.TestCase):
-    """Tests of the interface to Bandpass from lsst.sims.photUtils."""
-
-    def setUp(self):
-        """Define parameters used by every test."""
-        self.band_name = 'g'
-        self.wavelength_step = 10
-        self.bandpass = _load_bandpass(band_name=self.band_name, wavelength_step=self.wavelength_step)
-
-    def test_step_bandpass(self):
-        """Check that the bandpass has necessary methods, and those return the correct number of values."""
-        bp = self.bandpass
-        bp_wavelen, bandpass_vals = bp.getBandpass()
-        n_step = int(np.ceil((bp.wavelen_max - bp.wavelen_min) / bp.wavelen_step))
-        self.assertEqual(n_step + 1, len(bandpass_vals))
-
-
-class DcrModelTestBase(lsst.utils.tests.TestCase):
-
-    def setUp(self):
-        band_name = 'g'
-        n_step = 3
-        pixel_scale = 0.25
-        self.kernel_size = 5
-        self.size = 20
-        # NOTE that this array is randomly generated for each instance.
-        self.array = np.random.random(size=(self.size, self.size))
-        self.dcrModel = _BasicDcrModel(size=self.size, kernel_size=self.kernel_size, band_name=band_name,
-                                       n_step=n_step, pixel_scale=pixel_scale)
-        azimuth = 0.0
-        elevation = 70.0
-        self.dcr_gen = _dcr_generator(self.dcrModel.bandpass, pixel_scale=self.dcrModel.pixel_scale,
-                                      elevation=elevation, azimuth=azimuth, use_midpoint=False)
-        self.exposure = self.dcrModel._create_exposure(self.array, variance=None, elevation=elevation,
-                                                       azimuth=azimuth)
-
-    def tearDown(self):
-        del self.dcrModel
-        del self.exposure
-
-
-class KernelTestCase(DcrModelTestBase):
-    """Tests of the various kernels that incorporate dcr-based shifts."""
-
-    def test_simple_phase_kernel(self):
-        data_file = "test_data/simple_phase_kernel.npy"
-        psf = self.exposure.getPsf()
-        psf_size = psf.computeKernelImage().getArray().shape[0]
-        phase_arr = _calc_offset_phase(self.exposure, self.dcr_gen,
-                                       x_size=psf_size, y_size=psf_size, return_matrix=True)
-        phase_arr_ref = np.load(data_file)
-        self.assertFloatsEqual(phase_arr, phase_arr_ref)
-
-    def test_simple_psf_kernel(self):
-        data_file = "test_data/simple_psf_kernel.npy"
-        psf = self.exposure.getPsf()
-        psf_size = psf.computeKernelImage().getArray().shape[0]
-        phase_arr = _calc_psf_kernel(self.exposure, self.dcr_gen, x_size=psf_size, y_size=psf_size,
-                                     return_matrix=True, psf_img=self.dcrModel.psf_avg)
-        phase_arr_ref = np.load(data_file)
-        self.assertFloatsEqual(phase_arr, phase_arr_ref)
-
-    def test_full_psf_kernel(self):
-        data_file = "test_data/full_psf_kernel.npy"
-        psf = self.exposure.getPsf()
-        psf_size = psf.computeKernelImage().getArray().shape[0]
-        phase_arr = _calc_psf_kernel_full(self.exposure, self.dcr_gen, x_size=psf_size, y_size=psf_size,
-                                          return_matrix=True, psf_img=self.dcrModel.psf_avg)
-        phase_arr_ref = np.load(data_file)
-        self.assertFloatsEqual(phase_arr, phase_arr_ref)
-
-
-class DcrModelTestCase(DcrModelTestBase):
-    """Tests for the functions in the DcrModel class."""
-
-    def test_dataId_single(self):
-        id_ref = 100
-        band_ref = 'g'
-        ref_id = {'visit': id_ref, 'raft': '2,2', 'sensor': '1,1', 'filter': band_ref}
-        dataId = DcrModel._build_dataId(id_ref, band_ref)
-        self.assertEqual(ref_id, dataId[0])
-
-    def test_dataId_range(self):
-        id_ref = [100, 103]
-        band_ref = 'g'
-        ref_id = {'visit': id_ref, 'raft': '2,2', 'sensor': '1,1', 'filter': band_ref}
-        dataId = DcrModel._build_dataId(id_ref, band_ref)
-        for i, id in enumerate(range(id_ref[0], id_ref[1])):
-            ref_id = {'visit': id, 'raft': '2,2', 'sensor': '1,1', 'filter': band_ref}
-            self.assertEqual(ref_id, dataId[i])
-
-    def test_extract_model_no_weights(self):
-        # Make j and i different slightly so we can tell if the indices get swapped
-        i = self.size//2 + 1
-        j = self.size//2 - 1
-        radius = self.kernel_size//2
-        model_use = self.dcrModel.model
-        model_vals = DcrModel._extract_model_vals(j, i, radius=radius, model=model_use)
-        input_vals = [np.ravel(model_use[f, j - radius: j + radius + 1, i - radius: i + radius + 1])
-                      for f in range(self.dcrModel.n_step)]
-        self.assertFloatsEqual(np.hstack(input_vals), model_vals)
-
-    def test_extract_model_with_weights(self):
-        # Make j and i different slightly so we can tell if the indices get swapped
-        i = self.size//2 + 1
-        j = self.size//2 - 1
-        radius = self.kernel_size//2
-        model = self.dcrModel.model
-        weight_scale = 2.2
-        weights = self.dcrModel.weights * weight_scale
-        weights[:, j, i] = 0.
-        model_vals = DcrModel._extract_model_vals(j, i, radius=radius, model=model, weights=weights)
-        input_arr = []
-        for f in range(self.dcrModel.n_step):
-            input_vals = model[f, j - radius: j + radius + 1, i - radius: i + radius + 1] / weight_scale
-            input_vals[radius, radius] = 0.
-            input_arr.append(np.ravel(input_vals))
-
-        # input_vals = [model[_f, _j - radius: _j + radius + 1, _i - radius: _i + radius + 1] * weight_scale
-        #               for _f in range(self.dcrModel.n_step)]
-        # # input_vals = np.asarray(input_vals)
-        # input_vals[:][radius, radius] = 0.
-        # input_vals = [np.ravel(input_vals[_f]) for _f in range(self.dcrModel.n_step)]
-        self.assertFloatsEqual(np.hstack(input_arr), model_vals)
-
-    def test_apply_kernel(self):
-        data_file = "test_data/dcr_kernel_vals.npy"
-        i_use = self.size//2
-        j_use = self.size//2
-        radius = self.kernel_size//2
-        model_vals = DcrModel._extract_model_vals(j_use, i_use, radius=radius, model=self.dcrModel.model,
-                                                  weights=self.dcrModel.weights)
-        dcr_kernel = _calc_offset_phase(self.exposure, self.dcr_gen, return_matrix=True,
-                                        x_size=self.kernel_size, y_size=self.kernel_size)
-        dcr_vals = DcrModel._apply_dcr_kernel(dcr_kernel, model_vals, x_size=self.kernel_size,
-                                              y_size=self.kernel_size)
-        dcr_ref = np.load(data_file)
-        self.assertFloatsEqual(dcr_vals, dcr_ref)
-
-
-class PersistanceTestCase(DcrModelTestBase):
-    """Tests that read and write exposures and dcr models to disk."""
-
-    def test_create_exposure(self):
-        self.assertFloatsEqual(self.exposure.getMaskedImage().getImage().getArray(), self.array)
-        meta = self.exposure.getMetadata()
-        # Check that the required metadata is present:
-        try:
-            meta.get("ZENITH")
-        except Exception as e:
-            raise e
-        try:
-            meta.get("AZIMUTH")
-        except Exception as e:
-            raise e
-
-    def test_persist_dcr_model_roundtrip(self):
-        # Instantiating the butler takes several seconds, so all butler-related tests are condensed into one.
-        model_repository = "test_data"
-
-        # First test that the model values are not changed from what is expected
-
-        # The type "dcrModel" is read in as a 32 bit float, set in the lsst.obs.lsstSim.LsstSimMapper policy
-        model = np.float32(self.dcrModel.model)
-        self.dcrModel.export_model(model_repository=model_repository)
-        dcrModel2 = DcrModel(model_repository=model_repository)
-        # Note that butler.get() reads the FITS file in 32 bit precision.
-        self.assertFloatsEqual(model, dcrModel2.model)
-
-        # Next, test that the required parameters have been restored
-        param_ref = self.dcrModel.__dict__
-        param_new = dcrModel2.__dict__
-        for key in param_ref.keys():
-            self.assertIn(key, param_new)
-
-    def test_generate_template(self):
-        data_file = "test_data/template.npy"
-        elevation_arr = [50., 70., 85.]
-        az = 0.
-        # Note that self.array is randomly generated each call. That's okay, because the template should
-        # depend only on the metadata.
-        exposures = [self.dcrModel._create_exposure(self.array, variance=None, elevation=el, azimuth=az)
-                     for el in elevation_arr]
-        model_gen = self.dcrModel.generate_templates_from_model(exposures=exposures, kernel_size=5)
-        model_ref = np.load(data_file)
-        self.assertItemsEqual(model_gen, model_ref)
-
-
-class DcrModelGenerationTestCase(lsst.utils.tests.TestCase):
-
-    def setUp(self):
-        band_name = 'g'
-        self.n_step = 3
-        self.n_images = 5
-        pixel_scale = 0.25
-        self.kernel_size = 5
-        self.size = 20
-        use_psf = False
-
-        dcrModel = _BasicDcrModel(size=self.size, kernel_size=self.kernel_size, band_name=band_name,
-                                  n_step=self.n_step, pixel_scale=pixel_scale)
-
-        exposures = []
-        self.ref_vals = []
-        for i in range(self.n_images):
-            # NOTE that this array is randomly generated for each instance.
-            array = np.random.random(size=(self.size, self.size))*1000.
-            self.ref_vals.append(array)
-            el = np.random.random()*50. + 40.
-            az = np.random.random()*360.
-            exposures.append(dcrModel._create_exposure(array, variance=None, elevation=el, azimuth=az))
-        self.dcrCorr = _BasicDcrCorrection(kernel_size=self.kernel_size, band_name=band_name,
-                                           n_step=self.n_step, exposures=exposures, use_psf=use_psf)
-        self.dcrCorr.psf = dcrModel.psf
-        self.dcrCorr.psf_avg = dcrModel.psf_avg
-
-    def tearDown(self):
-        del self.dcrCorr
-
-    def test_extract_image(self):
-        # Make j and i different slightly so we can tell if the indices get swapped
-        i = self.size//2 + 1
-        j = self.size//2 - 1
-        radius = self.kernel_size//2
-        image_vals = self.dcrCorr._extract_image_vals(j, i, radius=radius)
-        input_vals = [np.ravel(self.ref_vals[f][j - radius: j + radius + 1, i - radius: i + radius + 1])
-                      for f in range(self.n_images)]
-        self.assertFloatsEqual(np.hstack(input_vals), image_vals)
-
-    def test_insert_model_vals(self):
-        i = self.size//2 + 1
-        j = self.size//2 - 1
-        radius = self.kernel_size//2
-        test_vals = np.random.random(size=(self.n_step, self.kernel_size, self.kernel_size))
-        model_ref = np.zeros((self.n_step, self.size, self.size))
-        self.dcrCorr.model = model_ref.copy()
-        weights_ref = np.zeros((self.n_step, self.size, self.size))
-        self.dcrCorr.weights = weights_ref.copy()
-        self.dcrCorr._insert_model_vals(j, i, test_vals, radius=radius)
-        psf_use = self.dcrCorr.psf_avg
-        model_ref[:, j - radius: j + radius + 1, i - radius: i + radius + 1] += test_vals*psf_use
-        weights_ref[:, j - radius: j + radius + 1, i - radius: i + radius + 1] += psf_use
-        self.assertFloatsEqual(model_ref, self.dcrCorr.model)
-        self.assertFloatsEqual(weights_ref, self.dcrCorr.weights)
-
-    def test_calculate_psf(self):
-        data_file = "test_data/calculate_psf.npy"
-        self.dcrCorr._calc_psf_model()
-        psf_size = self.dcrCorr.psf.computeKernelImage().getArray().shape[0]
-        p0 = psf_size//2 - self.kernel_size//2
-        p1 = p0 + self.kernel_size
-        psf_new = self.dcrCorr.psf.computeKernelImage().getArray()[p0: p1, p0: p1]
-        psf_ref = np.load(data_file)
-        self.assertFloatsEqual(psf_ref, psf_new)
-
-
-class RegularizationTestCase(lsst.utils.tests.TestCase):
-    def setUp(self):
-        self.kernel_size = 5
-        self.n_step = 3
-
-    def test_spatial_regularization(self):
-        data_file = "test_data/spatial_regularization.npy"
-        reg = DcrCorrection._build_regularization(x_size=self.kernel_size, y_size=self.kernel_size,
-                                                  n_step=self.n_step, spatial_regularization=True)
-        test_reg = np.load(data_file)
-        self.assertFloatsEqual(reg, test_reg)
-
-    def test_no_regularization(self):
-        # All regularization is set to False by default
-        reg = DcrCorrection._build_regularization(x_size=self.kernel_size, y_size=self.kernel_size,
-                                                  n_step=self.n_step)
-        self.assertIsNone(reg)
-
-    def test_frequency_regularization(self):
-        data_file = "test_data/frequency_regularization.npy"
-        reg = DcrCorrection._build_regularization(x_size=self.kernel_size, y_size=self.kernel_size,
-                                                  n_step=self.n_step, frequency_regularization=True)
-        test_reg = np.load(data_file)
-        self.assertFloatsEqual(reg, test_reg)
-
-    def test_frequency_derivative_regularization(self):
-        data_file = "test_data/frequency_derivative_regularization.npy"
-        reg = DcrCorrection._build_regularization(x_size=self.kernel_size, y_size=self.kernel_size,
-                                                  n_step=self.n_step, frequency_second_regularization=True)
-        test_reg = np.load(data_file)
-        self.assertFloatsEqual(reg, test_reg)
-
-    def test_multiple_regularization(self):
-        spatial_file = "test_data/spatial_regularization.npy"
-        freq_file = "test_data/frequency_regularization.npy"
-        deriv_file = "test_data/frequency_derivative_regularization.npy"
-
-        reg = DcrCorrection._build_regularization(x_size=self.kernel_size, y_size=self.kernel_size,
-                                                  n_step=self.n_step, spatial_regularization=True,
-                                                  frequency_regularization=True,
-                                                  frequency_second_regularization=True)
-        freq_reg = np.append(np.load(freq_file), np.load(deriv_file), axis=1)
-        test_reg = np.append(np.load(spatial_file), freq_reg, axis=1)
-        self.assertFloatsEqual(reg, test_reg)
-
-
-class SolverTestCast(lsst.utils.tests.TestCase):
-    def setUp(self):
-        data_file = "test_data/exposures.npy"
-        exposures = np.load(data_file)
-        self.dcrCorr = _BasicDcrCorrection(band_name='g', n_step=3, kernel_size=5, exposures=exposures)
-
-    def tearDown(self):
-        del self.dcrCorr
-
-    def test_build_dcr_kernel_full(self):
-        data_file = "test_data/build_dcr_kernel_full_vals.npy"
-        kernel = self.dcrCorr._build_dcr_kernel(use_full=True)
-        kernel_ref = np.load(data_file)
-        self.assertFloatsEqual(kernel, kernel_ref)
-
-    def test_build_dcr_kernel(self):
-        data_file = "test_data/build_dcr_kernel_vals.npy"
-        kernel = self.dcrCorr._build_dcr_kernel(use_full=False)
-        kernel_ref = np.load(data_file)
-        self.assertFloatsEqual(kernel, kernel_ref)
-
-    def test_build_model(self):
-        """Call build_model with as many options as possible turned off."""
-        data_file = "test_data/build_model_vals.npy"
-        self.dcrCorr.psf_avg = 1.
-        self.dcrCorr.build_model(use_full=False, use_regularization=False,
-                                 use_only_detected=False, verbose=False)
-        model_vals = self.dcrCorr.model
-        model_ref = np.load(data_file)
-        self.assertFloatsEqual(model_vals, model_ref)
-
-    def test_solve_model_no_regularization(self):
-        data_file = "test_data/solve_model_vals.npy"
-        y_size, x_size = self.dcrCorr.exposures[0].getDimensions()
-        pix_radius = self.dcrCorr.kernel_size//2
-        # Make j and i different slightly so we can tell if the indices get swapped
-        i = x_size//2 + 1
-        j = y_size//2 - 1
-        image_vals = self.dcrCorr._extract_image_vals(j, i, radius=pix_radius)
-        dcr_kernel = self.dcrCorr._build_dcr_kernel(use_full=False)
-        model_vals = self.dcrCorr._solve_model(dcr_kernel, image_vals, use_regularization=False)
-        model_ref = np.load(data_file)
-        self.assertFloatsEqual(model_vals, model_ref)
-
-    def test_solve_model_with_regularization(self):
-        data_file = "test_data/solve_model_reg_vals.npy"
-        y_size, x_size = self.dcrCorr.exposures[0].getDimensions()
-        kernel_size = self.dcrCorr.kernel_size
-        n_step = self.dcrCorr.n_step
-        pix_radius = kernel_size//2
-        i = x_size//2 + 1
-        j = y_size//2 - 1
-        image_vals = self.dcrCorr._extract_image_vals(j, i, radius=pix_radius)
-        dcr_kernel = self.dcrCorr._build_dcr_kernel(use_full=False)
-        self.dcrCorr.regularize = DcrCorrection._build_regularization(x_size=kernel_size, y_size=kernel_size,
-                                                                      n_step=n_step,
-                                                                      frequency_regularization=True)
-        model_vals = self.dcrCorr._solve_model(dcr_kernel, image_vals, use_regularization=True)
-        model_ref = np.load(data_file)
-        self.assertFloatsEqual(model_vals, model_ref)
-
-
-class MemoryTester(lsst.utils.tests.MemoryTestCase):
-    pass
-
-
-def setup_module(module):
-    lsst.utils.tests.init()
-
-
-if __name__ == "__main__":
-    lsst.utils.tests.init()
-    unittest.main()
